requirements.txt
africastalking==1.2.4
AfricastalkingGateway==2.0
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
contextlib2==21.6.0
Django==3.0
django-annoying==0.10.6
django-countries==7.3.2
django-crispy-forms==1.14.0
django-filter==21.1
django-google-translate==1.1
django-jsonstore==0.5.0
django-material==1.10.0
django-modeltranslation==0.18.2
django-pure-pagination==0.3.0
django-viewflow==1.11.0
idna==3.3
pytz==2022.1
requests==2.27.1
schema==0.7.5
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
